Refactors the cli-tool and add the Action service
