+++ sdk/tools/cli-tool/src/main/java/esa/mo/nmf/clitool/Helper.java
@@ -148,5 +148,5 @@ public class Helper {
             this.id = id;
         }
     }
-    
+
 }
