+++ sdk/tools/cli-tool/src/main/java/esa/mo/nmf/clitool/sm/SoftwareManagementCommands.java
@@ -18,9 +18,11 @@
  * limitations under the License.
  * ----------------------------------------------------------------------------
  */
-package esa.mo.nmf.clitool;
+package esa.mo.nmf.clitool.sm;
 
 import esa.mo.helpertools.connections.ConnectionConsumer;
+import esa.mo.nmf.clitool.BaseCommand;
+import esa.mo.nmf.clitool.ExitCodes;
 import static esa.mo.nmf.clitool.BaseCommand.consumer;
 import esa.mo.nmf.clitool.sm.AppsLauncherCommands.KillApp;
 import esa.mo.nmf.clitool.sm.AppsLauncherCommands.MonitorExecution;
